Conflicts:
	CMakeLists.txt

+++ fst/http/HttpHandler.hh
@@ -216,7 +216,7 @@ public:
   bool
   DecodeByteRange (std::string               rangeheader,
                    std::map<off_t, ssize_t> &offsetmap,
-                   ssize_t                  &requestsize,
+                   off_t                  &requestsize,
                    off_t                     filesize);
 
   /**
